third_party/blink/web_tests/external/wpt/web-animations/timing-model/animations/setting-the-target-effect-of-an-animation-expected.txt

This is a testharness.js-based test.
PASS If new effect is null and old effect is not null the animation becomes finish-pending
PASS If animation has a pending pause task, reschedule that task to run as soon as animation is ready.
PASS If animation has a pending play task, reschedule that task to run as soon as animation is ready to play new effect.
PASS The pending play task should be rescheduled even after temporarily setting the effect to null
FAIL When setting the effect of an animation to the effect of an existing animation, the existing animation's target effect should be set to null. assert_equals: expected "finished" but got "idle"
PASS After setting the target effect of animation to the target effect of an existing animation, the target effect's timing is updated to reflect the current time of the new animation.
PASS Setting the target effect to null causes a pending playback rate to be applied
Harness: the test ran to completion.
